+++ src/SWP/Bundle/CoreBundle/Service/ArticlePreviewTemplateHelperInterface.php
@@ -16,7 +16,7 @@ declare(strict_types=1);
 
 namespace SWP\Bundle\CoreBundle\Service;
 
-use SWP\Bundle\CoreBundle\Model\ArticleInterface;
+use SWP\Bundle\ContentBundle\Model\ArticleInterface;
 
 interface ArticlePreviewTemplateHelperInterface
 {