+++ src/SWP/Bundle/CoreBundle/Service/ArticlePreviewer.php
@@ -17,8 +17,8 @@ declare(strict_types=1);
 namespace SWP\Bundle\CoreBundle\Service;
 
 use SWP\Bundle\ContentBundle\Factory\ArticleFactoryInterface;
+use SWP\Bundle\ContentBundle\Model\ArticleInterface;
 use SWP\Bundle\ContentBundle\Model\RouteInterface;
-use SWP\Bundle\CoreBundle\Model\ArticleInterface;
 use SWP\Bundle\CoreBundle\Model\PackageInterface;
 use SWP\Bundle\CoreBundle\Processor\ArticleBodyProcessorInterface;
 use SWP\Component\Common\Exception\NotFoundHttpException;